Remove Linux Support
It is now working on Windows XP so there's no longer any need
for it to work on linux, or at least for the time being
there isn't.

If in the future there is such need then It will have
to be added back because right now it is too much of a
problem.