Separate App and in venue menus
various changes to separate the menus displayed when taking orders
depending ong whether the sale is for an in venue customer or for an app sale.

The database got a few fields added to accomodate the new data
so I will probably have to modify the code that checks the sales.