New appsData
Had to add new tables and columns to the database to separate a bit more
logically the data from the apps.

On buttons I had to modify the new contextual menu creator to make it
more flexible.

The database had a new table added to store the apps information.

DbIO was updated to implement the new tables

Session added the appID and configMenu variables throughout the Session
and AppSession classes to detect when and what type of app it is.

Next step is to add a new table to record the sales of the apps because
they can't use the same Folio.

I will probably just use the same folio as a base and then hash
it to get some random characters.