New Foliador
Modified the way sessions get ID's and Folios, appSessions now get a
different kind of folio.

Sessions now get an ID and a label, the ID is used for reference and the
label is what is exposed to the UI.

The buttons where modified to capture this data change in case they need
it in the future.