cleanup and improvements on folios
modified the foliador because it was not working as intended in some cases.
The "activeSession" was not working as intended because it was searching for
the ID and not the Label, so that got fixed.
To fix that I had to modify the buttons to read the correct data.

Also modified the collector and the db to store the data correctly.

removed a bit of code from the "deleteSession" function so now you can
delete sessions that are not the current session.
Not sure if that is something that will stay as is.

cleaned up a bit.