Receive cash from apps
Modified the payDialog to accept payments that are less than the total
of the order because some apps do that.

To make that work I modified the AppSession by removing the customer
segmentation radial buttons and placed a "CASH" toggle button in their
place.

This new functionality inverts the way the pay dialog functions by not
allowing you to register an order if you pay more than the total.

Also cleaned up some print statements and increased the version number.