V3.0.0 Database modification
Modified the appTickets database to explicitly state when an order is
paid in cash.

Was expecting to note the difference with a negative value in "Change",
but later realized that that only applies the orders which have a
discount, so I had to make it more explicit.

I increased the major version number because the database is no
longer compatible with the previous ones.